// src/models/types.ts
export type Event = typeof Events[keyof typeof Events];
// ...
export interface ThingDescription {
    id: string;
    name: string;
    title: string;
    '@context': string | string[];
    '@type': string | string[];
    href?: string;
    base?: string;
    description?: string;
    profile?: string | string[];
    selectedCapability: string;
    connected: boolean;
    properties: Record<string, PropertyDescription>,
    groupId?: string;
    iconData?: string;

    [key: string]: any;
}
// ...
export interface PropertyDescription {
    name: string;
    "@type"?: string;
    type?: string,
    maximum?: number;
    minimum?: number;
    multipleOf?: number;
    readOnly?: boolean;
    description: string;
    enum: any[];
    forms: any[];
    unit?: string;
    value: any;

    [key: string]: any;
}